+++ .idea/misc.xml
@@ -12,6 +12,7 @@
         <entry key="../../../../layout/compose-model-1646377583070.xml" value="0.1" />
         <entry key="../../../../layout/compose-model-1658254260460.xml" value="0.36359797297297297" />
         <entry key="../../../../layout/compose-model-1658257889000.xml" value="0.36359797297297297" />
+        <entry key="../../../../layout/compose-model-1658263377886.xml" value="0.1" />
       </map>
     </option>
   </component>
