+++ app/src/main/java/inc/combustion/example/devices/DevicesScreen.kt
@@ -36,21 +36,209 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.selection.selectable
 import androidx.compose.material.*
 import androidx.compose.runtime.*
+import androidx.compose.runtime.snapshots.SnapshotStateList
+import androidx.compose.runtime.snapshots.SnapshotStateMap
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import inc.combustion.example.R
-import inc.combustion.framework.service.DeviceManager
 import inc.combustion.engineering.ui.components.SingleSelectDialog
+import inc.combustion.example.AppScaffold
 import inc.combustion.example.AppState
-import inc.combustion.framework.service.ProbeID
-import inc.combustion.framework.service.ProbeColor
+import inc.combustion.example.theme.Combustion_Red
+import inc.combustion.framework.service.*
+
+/**
+ * State data object for a probe.  Binds the state between the DeviceScreen's ViewModel
+ * and Composable functions.  All of the properties in this class are observable using Kotlin State.
+ *
+ * @property serialNumber serial number.
+ * @property macAddress Bluetooth MAC address.
+ * @property firmwareVersion firmware version.
+ * @property hardwareRevision hardware revision.
+ * @property rssi current received signal strength.
+ * @property temperaturesCelsius current temperature values in Celsius.
+ * @property connectionState current connect state.
+ * @property units user's temperature units preference.
+ * @property uploadStatus user friendly status string of the upload
+ * @property recordsDownloaded number of records stored by the service
+ * @property recordRange the record range on the porbe
+ * @property color the probe's color setting.
+ * @property id the probes ID setting.
+ * @property instantRead the probe's Instant Read value.
+ */
+data class ProbeState(
+    val serialNumber: String,
+    var macAddress: MutableState<String> = mutableStateOf("TBD"),
+    var firmwareVersion: MutableState<String?> = mutableStateOf(null),
+    var hardwareRevision: MutableState<String?> = mutableStateOf(null),
+    var rssi: MutableState<Int> = mutableStateOf(0),
+    var temperaturesCelsius: SnapshotStateList<Double> = mutableStateListOf(
+        0.0,
+        0.0,
+        0.0,
+        0.0,
+        0.0,
+        0.0,
+        0.0,
+        0.0
+    ),
+    var connectionState: MutableState<ConnectionState> = mutableStateOf(ConnectionState.OUT_OF_RANGE),
+    var units: MutableState<Units> = mutableStateOf(Units.FAHRENHEIT),
+    var uploadStatus: MutableState<String> = mutableStateOf(""),
+    var recordsDownloaded: MutableState<Int> = mutableStateOf(0),
+    var recordRange: MutableState<String> = mutableStateOf(""),
+    var color: MutableState<String> = mutableStateOf(""),
+    var id: MutableState<String> = mutableStateOf(""),
+    var batteryStatus: MutableState<String> = mutableStateOf(""),
+    var instantRead: MutableState<String> = mutableStateOf("")
+) {
+    enum class Units(val string: String) {
+        FAHRENHEIT("Fahrenheit"),
+        CELSIUS("Celsius")
+    }
+
+    enum class ConnectionState(val string: String) {
+        OUT_OF_RANGE("Out of Range"),
+        ADVERTISING_CONNECTABLE("Advertising Connectable"),
+        ADVERTISING_NOT_CONNECTABLE("Advertising Not Connectable"),
+        CONNECTING("Connecting"),
+        CONNECTED("Connected"),
+        DISCONNECTING("Disconnecting"),
+        DISCONNECTED("Disconnected");
+
+        companion object {
+            fun fromDeviceConnectionState(state: DeviceConnectionState) : ConnectionState {
+                return when(state) {
+                    DeviceConnectionState.OUT_OF_RANGE -> OUT_OF_RANGE
+                    DeviceConnectionState.ADVERTISING_CONNECTABLE -> ADVERTISING_CONNECTABLE
+                    DeviceConnectionState.ADVERTISING_NOT_CONNECTABLE -> ADVERTISING_NOT_CONNECTABLE
+                    DeviceConnectionState.CONNECTING -> CONNECTING
+                    DeviceConnectionState.CONNECTED -> CONNECTED
+                    DeviceConnectionState.DISCONNECTING -> DISCONNECTING
+                    DeviceConnectionState.DISCONNECTED -> DISCONNECTED
+                }
+            }
+        }
+    }
+
+    enum class UploadState {
+        COMPLETE,
+        NEEDED
+    }
+
+    val T1 : MutableState<String> = mutableStateOf("")
+    val T2 : MutableState<String> = mutableStateOf("")
+    val T3 : MutableState<String> = mutableStateOf("")
+    val T4 : MutableState<String> = mutableStateOf("")
+    val T5 : MutableState<String> = mutableStateOf("")
+    val T6 : MutableState<String> = mutableStateOf("")
+    val T7 : MutableState<String> = mutableStateOf("")
+    val T8 : MutableState<String> = mutableStateOf("")
+
+    /**
+     * Updates this data object with the state update from the DeviceManager.
+     *
+     * @param state state update from the DeviceManager.
+     */
+    fun updateProbeState(state: Probe, downloads: Int) {
+        macAddress.value = state.mac
+        firmwareVersion.value = state.fwVersion
+        hardwareRevision.value = state.hwRevision
+        connectionState.value = ConnectionState.fromDeviceConnectionState(state.connectionState)
+        rssi.value = state.rssi
+        recordsDownloaded.value = downloads
+        color.value = state.color.toString()
+        id.value = state.id.toString()
+        batteryStatus.value = state.batteryStatus.toString()
+
+        instantRead.value = if(state.instantRead != null) {
+            String.format("%.1f", state.instantRead?.let { convertTemperature(it) })
+        } else {
+            "---"
+        }
+
+        if(state.temperatures != null) {
+            state.temperatures?.let {
+                T1.value = String.format("%.1f", convertTemperature(it.values[0]))
+                T2.value = String.format("%.1f", convertTemperature(it.values[1]))
+                T3.value = String.format("%.1f", convertTemperature(it.values[2]))
+                T4.value = String.format("%.1f", convertTemperature(it.values[3]))
+                T5.value = String.format("%.1f", convertTemperature(it.values[4]))
+                T6.value = String.format("%.1f", convertTemperature(it.values[5]))
+                T7.value = String.format("%.1f", convertTemperature(it.values[6]))
+                T8.value = String.format("%.1f", convertTemperature(it.values[7]))
+            }
+        } else {
+            T1.value = "---"
+            T2.value = "---"
+            T3.value = "---"
+            T4.value = "---"
+            T5.value = "---"
+            T6.value = "---"
+            T7.value = "---"
+            T8.value = "---"
+        }
+
+
+        uploadStatus.value = if(state.uploadState is ProbeUploadState.ProbeUploadInProgress) {
+            val inProgress = state.uploadState as ProbeUploadState.ProbeUploadInProgress
+            val percent = ((inProgress.recordsTransferred.toFloat() / inProgress.recordsRequested.toFloat()) * 100.0).toInt()
+
+            "$percent% of ${inProgress.recordsRequested}"
+        }
+        else if(state.uploadState is ProbeUploadState.ProbeUploadComplete){
+            UploadState.COMPLETE.toString()
+        }
+        else {
+            UploadState.NEEDED.toString()
+        }
+
+        recordRange.value = if(state.connectionState == DeviceConnectionState.CONNECTED)
+            "${state.minSequence} : ${state.maxSequence}"
+        else
+            ""
+    }
+
+    /**
+     * Converts the input temperature in Celsius to the user's current units preference
+     *
+     * @param temp Temperature in C
+     * @return temperature in preferred units.
+     */
+    private fun convertTemperature(temp: Double) : Double {
+        return if(units.value == Units.CELSIUS)
+            temp
+        else
+            (temp * 1.8) + 32.0
+    }
+}
+
+/**
+ * Data object for DeviceScreen state.
+ *
+ * @property probes Observable map of serial number -> ProbeState.
+ * @property onUnitsClick Lambda for handling click on units card button.
+ * @property onBluetoothClick Lambda for handling click on Bluetooth card button.
+ *
+ * @see ProbeState
+ * @see DevicesScreen
+ */
+data class DevicesScreenState(
+    val probes: SnapshotStateMap<String, ProbeState>,
+    val onUnitsClick: (ProbeState) -> Unit,
+    val onBluetoothClick: (ProbeState) -> Unit,
+    val onSetProbeColorClick: (String, ProbeColor) -> Unit,
+    val onSetProbeIDClick: (String, ProbeID) -> Unit
+)
 
 @Composable
 fun DevicesScreen(
@@ -76,13 +264,47 @@ fun DevicesScreen(
     )
 
     DevicesContent(
-        noDevicesReasonString = appState.noDevicesReasonString,
-        screenState = screenState
+        appState,
+        screenState
     )
 }
 
 @Composable
 fun DevicesContent(
+    appState: AppState,
+    screenState: DevicesScreenState
+) {
+    AppScaffold(
+        title = "Combustion Inc.",
+        navigationIcon = {
+            IconButton(onClick = { }) {
+                Icon(
+                    painter = painterResource(id = R.drawable.ic_launcher_foreground),
+                    contentDescription = null,
+                    tint = Combustion_Red
+                )
+            }
+        },
+        actionIcons = {
+            IconButton(onClick = appState.navigateToSettings()) {
+                Icon(
+                    tint = MaterialTheme.colors.onBackground,
+                    imageVector = ImageVector.vectorResource(id = R.drawable.ic_settings_24),
+                    contentDescription = null
+                )
+            }
+        },
+        appState = appState
+    ) {
+        DevicesList(
+            noDevicesReasonString = appState.noDevicesReasonString,
+            screenState = screenState
+        )
+    }
+}
+
+@Composable
+fun DevicesList(
     noDevicesReasonString: String,
     screenState: DevicesScreenState
 ) {
